Add gear_lock mapflag and raise MAP_MAX_MSG to 5000

+++ src/map/clif.cpp
@@ -10894,7 +10894,7 @@ void clif_parse_LoadEndAck(int32 fd,map_session_data *sd)
 		clif_spawn(sd->ed);
 		clif_elemental_info(sd);
 		clif_elemental_updatestatus(*sd, SP_HP);
-		clif_hpmeter_single( *sd, sd->ed->id, sd->ed->battle_status.hp, sd->ed->battle_status.max_hp );
+		clif_hpmeter_single( *sd, static_cast<uint32>( sd->ed->id ), static_cast<uint32>( sd->ed->battle_status.hp ), static_cast<uint32>( sd->ed->battle_status.max_hp ) );
 		clif_elemental_updatestatus(*sd, SP_SP);
 		status_calc_bl(sd->ed, { SCB_SPEED }); //Elemental mimic their master's speed on each map change
 	}
